pyproject.toml
[project]
name = "cours-mdp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "google-cloud-firestore>=2.23.0",
    "gunicorn>=25.0.3",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.15.0",
]
